Reject multiple calendar updates on new glucose data (fixes #276)

+++ FreeAPS/Sources/Services/Calendar/CalendarManager.swift
@@ -21,6 +21,7 @@ final class BaseCalendarManager: CalendarManager, Injectable {
 
     private var coreDataPublisher: AnyPublisher<Set<NSManagedObject>, Never>?
     private var subscriptions = Set<AnyCancellable>()
+    private var previousDeterminationId: NSManagedObjectID?
 
     private var glucoseFormatter: NumberFormatter {
         let formatter = NumberFormatter()
@@ -211,6 +212,11 @@ final class BaseCalendarManager: CalendarManager, Injectable {
         guard settingsManager.settings.useCalendar, let calendar = currentCalendar,
               let determinationId = await getLastDetermination() else { return }
 
+        // Ignore the update if the determinationId is the same as it was at last update
+        if determinationId == previousDeterminationId {
+            return
+        }
+
         let glucoseIds = await fetchGlucose()
 
         deleteAllEvents(in: calendar)
@@ -290,6 +296,8 @@ final class BaseCalendarManager: CalendarManager, Injectable {
             event.calendar = calendar
 
             try eventStore.save(event, span: .thisEvent)
+
+            previousDeterminationId = determinationId
 
         } catch {
             debugPrint(
